More atomic peephole code.
